moving tests to seperate files
