+++ exposed-tests/src/test/kotlin/org/jetbrains/exposed/sql/tests/shared/DDLTests.kt
@@ -840,61 +840,10 @@ class DDLTests : DatabaseTestsBase() {
         }
     }
 
-    // https://github.com/JetBrains/Exposed/issues/709
-    @Test
-    fun createTableWithDuplicateColumn() {
-        val assertionFailureMessage = "Can't create a table with multiple columns having the same name"
-
-        withDb() {
-            assertFails(assertionFailureMessage) {
-                SchemaUtils.create(TableWithDuplicatedColumn)
-            }
-            assertFails(assertionFailureMessage) {
-                SchemaUtils.create(TableDuplicatedColumnRefereToIntIdTable)
-            }
-            assertFails(assertionFailureMessage) {
-                SchemaUtils.create(TableDuplicatedColumnRefereToTable)
-            }
-        }
-    }
-
-    // https://github.com/JetBrains/Exposed/issues/709
-    @Test
-    fun replaceColumnToDuplicateColumn() {
-        val assertionFailureMessage = "Can't replace a column with another one that has the same name as an existing column"
-
-        withTables(IDTable) {
-            assertFailsWith(exceptionClass =    DuplicateColumnException::class,
-                            message =           assertionFailureMessage
-            ) {
-                //Duplicate the id column by replacing the IDTable.code by a column with the name "id"
-                val id = Column<Int>(IDTable, IDTable.id.name, IDTable.id.columnType)
-                IDTable.replaceColumn(IDTable.code, id)
-            }
-        }
-    }
-
     object TableFromSchemeOne : IntIdTable("one.test")
 
     object TableFromSchemeTwo : IntIdTable("two.test") {
         val reference = reference("testOne", TableFromSchemeOne)
-    }
-
-    object TableWithDuplicatedColumn : Table("myTable") {
-        val id1 = integer("id")
-        val id2 = integer("id")
-    }
-
-    object IDTable : IntIdTable("myTable") {
-        val code = integer("code")
-    }
-
-    object TableDuplicatedColumnRefereToIntIdTable : IntIdTable("myTable") {
-        val reference = reference("id", IDTable)
-    }
-
-    object TableDuplicatedColumnRefereToTable : Table("myTable") {
-        val reference = reference("id", TableWithDuplicatedColumn.id1)
     }
 }
